# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hacker Dashboard API",
    "description": "임의 투자 데이터 스키마에서 자동 분석 뷰를 생성하는 금융 대시보드 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/portfolio/holdings": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "List Holdings",
        "description": "보유 종목 전체 조회.",
        "operationId": "list_holdings_portfolio_holdings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HoldingResponse"
                  },
                  "type": "array",
                  "title": "Response List Holdings Portfolio Holdings Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Create Holding",
        "description": "보유 종목 추가.",
        "operationId": "create_holding_portfolio_holdings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/holdings/{holding_id}": {
      "patch": {
        "tags": [
          "portfolio"
        ],
        "summary": "Update Holding",
        "description": "수량/평단가 수정.",
        "operationId": "update_holding_portfolio_holdings__holding_id__patch",
        "parameters": [
          {
            "name": "holding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Holding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolio"
        ],
        "summary": "Delete Holding",
        "description": "보유 종목 삭제.",
        "operationId": "delete_holding_portfolio_holdings__holding_id__delete",
        "parameters": [
          {
            "name": "holding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Holding Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/portfolio/snapshots": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "List Snapshots",
        "description": "포트폴리오 스냅샷 조회 (날짜 범위 필터).",
        "operationId": "list_snapshots_portfolio_snapshots_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "시작 날짜 (YYYY-MM-DD)",
              "title": "From"
            },
            "description": "시작 날짜 (YYYY-MM-DD)"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "종료 날짜 (YYYY-MM-DD)",
              "title": "To"
            },
            "description": "종료 날짜 (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SnapshotResponse"
                  },
                  "title": "Response List Snapshots Portfolio Snapshots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/rebalance": {
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Rebalance",
        "description": "리밸런싱 제안 생성.\n\n결정적 계산(순수함수)으로 actions 배열을 생성하고,\nLLM 해석(RebalanceAnalyzer)을 옵션으로 추가.\nLLM 실패 시 llm_analysis=None, status=\"degraded\" 로 graceful degrade.",
        "operationId": "rebalance_portfolio_rebalance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RebalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldingCreate": {
        "properties": {
          "market": {
            "type": "string",
            "title": "Market",
            "description": "upbit | binance | yahoo | naver_kr"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "심볼/코드 (예: KRW-BTC, AAPL)"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantity",
            "description": "보유 수량"
          },
          "avg_cost": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Avg Cost",
            "description": "평균 매입 단가"
          },
          "currency": {
            "type": "string",
            "maxLength": 4,
            "minLength": 3,
            "title": "Currency",
            "description": "통화 코드 (KRW/USD/USDT/EUR/JPY)"
          }
        },
        "type": "object",
        "required": [
          "market",
          "code",
          "quantity",
          "avg_cost",
          "currency"
        ],
        "title": "HoldingCreate"
      },
...
      "HoldingResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "quantity": {
            "type": "string",
            "title": "Quantity"
          },
          "avg_cost": {
            "type": "string",
            "title": "Avg Cost"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "market",
          "code",
          "quantity",
          "avg_cost",
          "currency",
          "created_at",
          "updated_at"
        ],
        "title": "HoldingResponse"
      },
      "HoldingUpdate": {
        "properties": {
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "avg_cost": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cost"
          }
        },
        "type": "object",
        "title": "HoldingUpdate"
      },
      "LLMAnalysis": {
        "properties": {
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "narrative": {
            "type": "string",
            "title": "Narrative"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "headline",
          "narrative"
        ],
        "title": "LLMAnalysis"
      },
...
      "RebalanceAction": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Action"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "asset_class": {
            "type": "string",
            "title": "Asset Class"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          },
          "estimated_value_krw": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Value Krw"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "action",
          "market",
          "code",
          "asset_class",
          "quantity",
          "reason"
        ],
        "title": "RebalanceAction"
      },
      "RebalanceConstraints": {
        "properties": {
          "max_single_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Single Weight",
            "description": "단일 종목 최대 비중",
            "default": 0.5
          },
          "min_trade_krw": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Min Trade Krw",
            "description": "최소 거래액(KRW)",
            "default": "100000"
          },
          "allow_fractional": {
            "type": "boolean",
            "title": "Allow Fractional",
            "description": "소수점 수량 허용 (주식은 기본 False 권장)",
            "default": true
          }
        },
        "type": "object",
        "title": "RebalanceConstraints"
      },
      "RebalanceMeta": {
        "properties": {
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "gates": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Gates"
          },
          "evidence_snippets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Snippets"
          }
        },
        "type": "object",
        "required": [
          "gates"
        ],
        "title": "RebalanceMeta"
      },
      "RebalanceRequest": {
        "properties": {
          "target_allocation": {
            "$ref": "#/components/schemas/TargetAllocation"
          },
          "constraints": {
            "$ref": "#/components/schemas/RebalanceConstraints"
          }
        },
        "type": "object",
        "required": [
          "target_allocation"
        ],
        "title": "RebalanceRequest"
      },
      "RebalanceResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded",
              "error"
            ],
            "title": "Status"
          },
          "current_allocation": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Current Allocation"
          },
          "target_allocation": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Target Allocation"
          },
          "drift": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Drift"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/RebalanceAction"
            },
            "type": "array",
            "title": "Actions"
          },
          "expected_allocation": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Expected Allocation"
          },
          "summary": {
            "$ref": "#/components/schemas/RebalanceSummary"
          },
          "llm_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/RebalanceMeta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "status",
          "current_allocation",
          "target_allocation",
          "drift",
          "actions",
          "expected_allocation",
          "summary",
          "meta"
        ],
        "title": "RebalanceResponse"
      },
      "RebalanceSummary": {
        "properties": {
          "total_trades": {
            "type": "integer",
            "title": "Total Trades"
          },
          "total_sell_value_krw": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Sell Value Krw"
          },
          "total_buy_value_krw": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Buy Value Krw"
          },
          "rebalance_cost_estimate_krw": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Rebalance Cost Estimate Krw"
          }
        },
        "type": "object",
        "required": [
          "total_trades",
          "total_sell_value_krw",
          "total_buy_value_krw",
          "rebalance_cost_estimate_krw"
        ],
        "title": "RebalanceSummary"
      },
...
      "SnapshotResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date"
          },
          "total_value_krw": {
            "type": "string",
            "title": "Total Value Krw"
          },
          "total_pnl_krw": {
            "type": "string",
            "title": "Total Pnl Krw"
          },
          "asset_class_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Asset Class Breakdown"
          },
          "holdings_detail": {
            "items": {},
            "type": "array",
            "title": "Holdings Detail"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "snapshot_date",
          "total_value_krw",
          "total_pnl_krw",
          "asset_class_breakdown",
          "holdings_detail",
          "created_at"
        ],
        "title": "SnapshotResponse"
      },
...
      "TargetAllocation": {
        "properties": {
          "stock_kr": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Stock Kr",
            "default": 0.0
          },
          "stock_us": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Stock Us",
            "default": 0.0
          },
          "crypto": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Crypto",
            "default": 0.0
          },
          "cash": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Cash",
            "default": 0.0
          },
          "fx": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fx",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "TargetAllocation",
        "description": "자산군별 목표 비중. 합계 1.0 (±0.001 허용)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}